made getName( ) and getQuestion( )
The code from before is now compressed into 2 functions.
So, getName( ) gets the name of the user and ALSO, asks if their first word is a greeting.
eg.
userName = getName(<user's input>)

getQuestion might need to be looked over. Just call it to get the questions I guess ... It also tells the user if they got the question right and shows the answer. :]