Bug Fix: Symbol Equations
Text recognition will now recognise and solve symbol equations (Only tested on "1 + 1" currently) as EOF error has been solved by correcting i = char to char = i on line 36 as this now allows the if statement to check through the sentence properly instead of i (the element of the string being tested) being overwritten by the previously assigned value of char which was "" (blank) before the if statement could check it. Word equations are still not functioning correctly however.